+++ src/main.ts
@@ -1,3 +1,4 @@
+import { prepareActionData, prepareSkillData } from "@module/apps/helpers.ts";
 import { GMDialog, RollDialog } from "@module/apps/index.ts";
 import type { SocketRollRequest } from "@module/apps/types.ts";
 import * as R from "remeda";
@@ -20,8 +21,16 @@ Hooks.once("ready", () => {
     game.socket.on("module.pf2e-request-rolls", (request: SocketRollRequest, userId: string) => {
         const sender = game.users.get(userId, { strict: true });
         if (!sender.isGM) return;
-        new RollDialog({ request }).render({ force: true });
+
+        if (request.users.includes(game.user.id)) {
+            new RollDialog({ request }).render({ force: true });
+        }
     });
+});
+
+Hooks.once("pf2e.systemReady", () => {
+    prepareActionData();
+    prepareSkillData();
 });
 
 Hooks.on("getChatLogEntryContext", (_chatlog, options) => {